fix: pull resolution modules into separate file
