+++ packages/test-apps/react-vite/src/components/NeatPotionShelf.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { Asset, Entry, UnresolvedLink } from 'contentful';
-import { inMemoryEntities } from '@contentful/experiences-sdk-react';
+import { Asset, Entry } from 'contentful';
+import { copyAndResolveEntityToLevel3, isEntry, isLink } from '../resolutionUtilities';
 
 type ShallowEntry = Entry;
 type NeatPotionShelfProps = {
@@ -26,80 +26,6 @@ type Ingredient = {
   fields: {
     title: string;
   };
-};
-
-const isLink = (value: unknown): value is UnresolvedLink<'Asset'> | UnresolvedLink<'Entry'> => {
-  return (
-    null !== value &&
-    typeof value === 'object' &&
-    'sys' in value &&
-    (value as UnresolvedLink<'Asset'> | UnresolvedLink<'Entry'>).sys?.type === 'Link'
-  );
-};
-
-const isEntry = (value: unknown): value is Entry => {
-  return (
-    null !== value &&
-    typeof value === 'object' &&
-    'sys' in value &&
-    (value as Entry).sys?.type === 'Entry'
-  );
-};
-
-const isAsset = (value: unknown): value is Asset => {
-  return (
-    null !== value &&
-    typeof value === 'object' &&
-    'sys' in value &&
-    (value as Asset).sys?.type === 'Asset'
-  );
-};
-
-// TODO: how does it classify empty arrays?
-const isArrayOfLinks = (
-  value: unknown,
-): value is Array<UnresolvedLink<'Asset'> | UnresolvedLink<'Entry'>> => {
-  return Array.isArray(value) && value.every((item) => isLink(item));
-};
-
-const resolveLinkOrArrayOrPassthrough = (
-  fieldValue: unknown | UnresolvedLink<'Asset'> | UnresolvedLink<'Entry'>,
-) => {
-  if (isLink(fieldValue)) {
-    return inMemoryEntities.maybeResolveLink(
-      fieldValue as unknown as UnresolvedLink<'Asset'> | UnresolvedLink<'Entry'>,
-    );
-  }
-
-  if (isArrayOfLinks(fieldValue)) {
-    return fieldValue.map((link) => inMemoryEntities.maybeResolveLink(link));
-  }
-
-  // we just pass through the value
-  return fieldValue;
-};
-
-// TODO: maybe we can return some kind of calculation, as to - how many links failed to resolve?
-const copyAndResolveEntityToLevel3 = (
-  entity: Entry | Asset | undefined,
-): Entry | Asset | undefined => {
-  if (entity === undefined) {
-    return undefined;
-  }
-  if (isAsset(entity)) {
-    return structuredClone(entity);
-  }
-  if (isEntry(entity)) {
-    const e = structuredClone(entity) as Entry;
-    for (const [fname, value] of Object.entries(e.fields)) {
-      // @ts-expect-error casting unkonwn
-      e.fields[fname] = resolveLinkOrArrayOrPassthrough(value); // default behaviour when link is not resolved is to return undefined
-    }
-    return e;
-  }
-
-  // everything else just pass through, but copy
-  return structuredClone(entity);
 };
 
 export const NeatPotionShelf: React.FC<NeatPotionShelfProps> = (props: NeatPotionShelfProps) => {
